pipelines/google_adwords/requirements.txt
pandas
pendulum
wheel
google-ads
snowflake
